added NewProjectModal component
This modal will be where the user fills out the form for a new project.

I took the demo code from the material-ui documentation and slightly
altered it to get it to work on the Projects page. I have yet to add
any of the actual form inputs and styling.